Fix notification trigger bug
Fixed bug that would make notifications not show up. This was due
to the notifications being created with the same id. Meaning that
they would activate once on a clean build and never repeat (because
the notification still exists).

Move NotificationManager and NotificationCenterDelegate to separate
files. Added some logging to each file.